split util func, moved concepts with similar function to utils
