+++ python/aristotle-metadata-registry/aristotle_mdr/contrib/slots/models.py
@@ -37,27 +37,3 @@ class Slot(TimeStampedModel):
         ordering = ['order']
 
 
-def concepts_with_similar_slots(user, name=None, _type=None, value=None, slot=None):
-    if not (slot is not None or _type is not None or name is not None):
-        return MDR._concept.objects.none()
-
-    if slot is not None:
-        name = slot.name
-        _type = slot.type
-        value = slot.value
-
-    slots = MDR._concept.objects.visible(user)
-
-    if name:
-        slots = slots.filter(slots__name=name)
-
-    if _type:
-        slots = slots.filter(slots__type=_type)
-
-    if value:
-        slots = slots.filter(slots__value=value)
-
-    if slot:
-        slots = slots.exclude(id=slot.concept.id)
-
-    return slots.distinct()
