+++ python/aristotle-metadata-registry/aristotle_mdr/contrib/slots/templatetags/slot_tags.py
@@ -1,6 +1,7 @@
 from django import template
 
-from aristotle_mdr.contrib.slots.models import Slot, concepts_with_similar_slots
+from aristotle_mdr.contrib.slots.models import Slot
+from aristotle_mdr.contrib.slots.utils import concepts_with_similar_slots
 
 
 register = template.Library()
