+++ python/aristotle-metadata-registry/aristotle_mdr/contrib/slots/utils.py
@@ -1,9 +1,7 @@
 from aristotle_mdr.contrib.slots.models import Slot
+from aristotle_mdr import models as MDR
 
-
-def get_allowed_slots(concept, user):
-
-    slots = Slot.objects.filter(concept=concept)
+def filter_slot_perms(slots, user):
 
     # If slots is empty no need for further filtering
     if not slots:
@@ -19,3 +17,33 @@ def get_allowed_slots(concept, user):
     else:
         # Only return public slots
         return slots.filter(permission=0)
+
+def get_allowed_slots(concept, user):
+
+    slots = Slot.objects.filter(concept=concept)
+    return filter_slot_perms(slots, user)
+
+def concepts_with_similar_slots(user, name=None, _type=None, value=None, slot=None):
+    if not (slot is not None or _type is not None or name is not None):
+        return MDR._concept.objects.none()
+
+    if slot is not None:
+        name = slot.name
+        _type = slot.type
+        value = slot.value
+
+    slots = MDR._concept.objects.visible(user)
+
+    if name:
+        slots = slots.filter(slots__name=name)
+
+    if _type:
+        slots = slots.filter(slots__type=_type)
+
+    if value:
+        slots = slots.filter(slots__value=value)
+
+    if slot:
+        slots = slots.exclude(id=slot.concept.id)
+
+    return slots.distinct()
