+++ python/aristotle-metadata-registry/aristotle_mdr/contrib/slots/views.py
@@ -1,7 +1,8 @@
 from django.views.generic import ListView
 
 from aristotle_mdr.models import _concept
-from aristotle_mdr.contrib.slots.models import Slot, concepts_with_similar_slots
+from aristotle_mdr.contrib.slots.models import Slot
+from aristotle_mdr.contrib.slots.utils import concepts_with_similar_slots
 
 
 class SimilarSlotsView(ListView):
